Crate bitfrob

A crate to help with bit manipulation of integers.

Structs

U8ConstRegionMask

Like u8_region_mask, but forces the value into an associated constant.

U16ConstRegionMask

Like u16_region_mask, but forces the value into an associated constant.

U32ConstRegionMask

Like u32_region_mask, but forces the value into an associated constant.

u8x2

Two u8 values packed as a u16.

Functions

u8_get_bit

Gets the B bit.

u8_get_region

Gets a L to H (inclusive) bit region of the value.

u8_get_value

Like u8_get_region but the output is shifted down appropriately.

u8_region_mask

Generates a bit mask where all bits in the region are 1.

u8_with_bit

Replaces the B bit.

u8_with_region

Replaces a L to H (inclusive) bit region of the value.

u8_with_value

Like u8_with_region but the value is shifted up appropriately.

u16_get_bit

Gets the B bit.

u16_get_region

Gets a L to H (inclusive) bit region of the value.

u16_get_value

Like u16_get_region but the output is shifted down appropriately.

u16_region_mask

Generates a bit mask where all bits in the region are 1.

u16_with_bit

Replaces the B bit.

u16_with_region

Replaces a L to H (inclusive) bit region of the value.

u16_with_value

Like u16_with_region but the value is shifted up appropriately.

u32_get_bit

Gets the B bit.

u32_get_region

Gets a L to H (inclusive) bit region of the value.

u32_get_value

Like u32_get_region but the output is shifted down appropriately.

u32_region_mask

Generates a bit mask where all bits in the region are 1.

u32_with_bit

Replaces the B bit.

u32_with_region

Replaces a L to H (inclusive) bit region of the value.

u32_with_value

Like u32_with_region but the value is shifted up appropriately.