Crate bitfrob

A crate to help with bit manipulation of integers.

Naming Conventions

The bit function work on a single bit.

The region functions work on multiple adjacent bits.

The value functions are like a region but the value is automatically shifted down on read and up on write so that you can input/output normal values and the function will put it in place for you.

Structs

u8x2

Two u8 values packed as a u16.

Functions

u8_get_bit

Determines if the b bit is set in u.

u8_get_region

Get the low to high bit region of u.

u8_get_value

Get the low to high bit region of u shifted to be a normal value.

u8_region_mask

Generates a bit mask where all bits in the region are 1.

u8_with_bit

Replaces the b bit in u, returning the new value.

u8_with_region

Replaces the low to high bit region of u_old, returning the new value.

u8_with_value

Replaces the low to high bit region of u_old, returning the new value.

u16_get_bit

Determines if the b bit is set in u.

u16_get_region

Get the low to high bit region of u.

u16_get_value

Get the low to high bit region of u shifted to be a normal value.

u16_region_mask

Generates a bit mask where all bits in the region are 1.

u16_with_bit

Replaces the b bit in u, returning the new value.

u16_with_region

Replaces the low to high bit region of u_old, returning the new value.

u16_with_value

Replaces the low to high bit region of u_old, returning the new value.

u32_get_bit

Determines if the b bit is set in u.

u32_get_region

Get the low to high bit region of u.

u32_get_value

Get the low to high bit region of u shifted to be a normal value.

u32_region_mask

Generates a bit mask where all bits in the region are 1.

u32_with_bit

Replaces the b bit in u, returning the new value.

u32_with_region

Replaces the low to high bit region of u_old, returning the new value.

u32_with_value

Replaces the low to high bit region of u_old, returning the new value.

u64_get_bit

Determines if the b bit is set in u.

u64_get_region

Get the low to high bit region of u.

u64_get_value

Get the low to high bit region of u shifted to be a normal value.

u64_region_mask

Generates a bit mask where all bits in the region are 1.

u64_with_bit

Replaces the b bit in u, returning the new value.

u64_with_region

Replaces the low to high bit region of u_old, returning the new value.

u64_with_value

Replaces the low to high bit region of u_old, returning the new value.

u128_get_bit

Determines if the b bit is set in u.

u128_get_region

Get the low to high bit region of u.

u128_get_value

Get the low to high bit region of u shifted to be a normal value.

u128_region_mask

Generates a bit mask where all bits in the region are 1.

u128_with_bit

Replaces the b bit in u, returning the new value.

u128_with_region

Replaces the low to high bit region of u_old, returning the new value.

u128_with_value

Replaces the low to high bit region of u_old, returning the new value.